Introduce CPP to the build
Switch around the import file to use C++ so that I can take advantage
of the ease of string processing that comes with that.
